requirements.txt
gdown
tensorflow
streamlit
Pillow
numpy